project/cmdb/requirements.txt
Django==2.0.4
django-auth-ldap==1.3.0
django-crontab==0.7.1
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.8.1
ldap3==2.4
Markdown==2.6.11
paramiko==2.4.1
Pygments==2.2.0
pyldap==3.0.0.post1
PyMySQL==0.8.0
python-ldap==3.1.0
requests==2.19.1